+++ jsonscribe/utils.py
@@ -42,3 +42,23 @@ def utcnow():
 
     """
     return datetime.datetime.now(tz=utc)
+
+
+def parse_boolean(s):
+    """Parse a string as a boolean value.
+
+    :returns: boolean value that `s` matches
+    :raises: :exc:`ValueError` if `s` does not match a known value
+
+    Acceptable values are 1, y, yes, t, or true for *truthiness* and
+    0, n, no, f, or false for *falsiness*.
+
+    """
+    if s is None:
+        s = ''
+    value = s.strip().lower()
+    if value in ('0', 'n', 'no', 'f', 'false'):
+        return False
+    if value in ('1', 'y', 'yes', 't', 'true'):
+        return True
+    raise ValueError('cannot determine Boolean value: %r' % (s, ))